pyproject.toml
[tool.poetry]
name = "same-installer-website"
version = "0.1.0"
description = ""
authors = ["David Aronchick <aronchick@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Flask = "^1.1.2"
PyGithub = "^1.54.1"
ipython = "^7.21.0"
regex = "^2021.3.17"

[tool.poetry.dev-dependencies]
black = {version = "^20.8b1", allow-prereleases = true}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
